Fix sleeping bug in auth server registration
1) Call of med_cache_alloc_size() from l4_add_kclass() under
   registry_lock caused the bug. Fixed by changing GFP_KERNEL to
   GFP_NOWAIT in kmem_cache_alloc() call in kmem_cache_alloc_index().
2) Registration of memory cache in l4_add_kclass() caused another bug.
   This cannot be done under lock, so it was moved to L3. This caused a
   new problem:
3) As memory cache array was allocated in user_open(), it doesn't exist
   during registration of all kclasses and kevents in L3. The
   allocation was moved to fs_initcall, so it is done before
   initialization of L2 objects through device_initcall.
